refactor: :recycle: saque las funciones de app y las pase al archivo "functions.js"
