+++ api/src/app.js
@@ -1,37 +1,8 @@
 const express = require('express')
-const multer = require('multer')
 const fs = require('fs')
 const path = require('path')
-
+const { upload } = require('./functions')
 const server = express()
-
-// Configurar el almacenamiento y filtrar por tipo de archivo
-const storage = multer.diskStorage({
-  destination: path.join(__dirname, '../Test/download'),
-  filename: (req, file, cb) => {
-    // Define el nombre del archivo en el servidor
-    cb(null, file.originalname)
-  }
-})
-
-const fileFilter = (req, file, cb) => {
-  // Filtrar por tipo de archivo permitido (en este caso, solo imágenes)
-  const typeFilter = /jpeg|jpg|gif/
-  const array = file.originalname.split('.')
-  const isValid = array[1]
-  if (typeFilter.test(isValid)) {
-    cb(null, true) // Aceptar el archivo
-  } else {
-    cb(new Error('Tipo de archivo no permitido'), false) // Rechazar el archivo
-  }
-}
-
-// Configurar Multer con el almacenamiento y el filtro de archivos
-const upload = multer({
-  storage,
-  fileFilter,
-  limits: { fileSize: 102400 } // Limite de tamaño en bytes (100 KB)
-})
 
 // Manejador de la ruta que utiliza Multer para cargar archivos
 server.post('/upload', upload.single('archivo'), (req, res) => {
